refactor: address PR review - PascalCase, useCliCommandRunner hook, IPC exit
